Basic minimal comments and formatting.
Basically just applied formatting consistently throughout and organized the config file a bit. Also tried to quickly add some minimal code comments. Finally, the graph.txt was changed to be written to a temp directory, but it was still in the repo and that's where it was being read from it seems. Fixed that as well.